chore: cleanup infra files into helpers
jsut trying to reduce the noise on what is reusable or not
